+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Chg #159: Replace `withNextPageToken()` and `withPreviousPageToken()` of `PaginatorInterface` with `withToken()`,
   `getNextPageToken()` with `getNextToken()`, `getPreviousPageToken()` with `getPreviousToken()`, and add `getToken()`.
   These methods use new `PageToken` class (@vjik)
+- New #160: Add `Sort::hasFieldInConfig()` method that checks for the presence of a field in the sort config (@vjik)
 
 ## 1.0.1 January 25, 2023